[add]Policyでのアクセス制御

+++ app/Http/Controllers/PostController.php
@@ -57,6 +57,7 @@ class PostController extends Controller
 
     public function edit(Post $post)
     {
+        $this->authorize('update', $post);
         return view('post.edit', compact('post'));
     }
 
@@ -74,7 +75,8 @@ class PostController extends Controller
     }
 
     public function destroy(Post $post)
-    {
+    {   
+        $this->authorize('delete', $post);
         $this->postRepository->delete($post);
 
         return redirect()->route('post.index')->with('message', '投稿を削除しました');
